examples/trading-assistant-langgraph/requirements.txt
langgraph
langchain-core
langchain-community
langchain-ollama
yfinance
pandas
python-dotenv 